Refactoring to adopt best practices
