+++ .env.example
@@ -1,0 +1,5 @@
+MYSQL_HOST=example
+MYSQL_USER=example
+MYSQL_PASSWORD=example
+MYSQL_DB=example
+SECRET_KEY=secret_key
