+++ CHEMIST/data.py
@@ -1,0 +1,17 @@
+# Dummy data for products
+products_data = {
+    1: {"name": "Product 1", "price": 900, "image_url": "/static/images/product1.png"},
+    2: {"name": "Product 2", "price": 300, "image_url": "/static/images/product2.png"},
+    3: {"name": "Product 3", "price": 300, "image_url": "/static/images/product3.png"}
+}
+
+# Mock sales data
+sales_data = {
+    "2024-05-17": [
+        {"date": "2024-05-17", "product_name": "Medicine A", "quantity": 10, "total_price": 5000},
+        {"date": "2024-05-17", "product_name": "Medicine B", "quantity": 5, "total_price": 2500}
+    ],
+    "2024-05-18": [
+        {"date": "2024-05-18", "product_name": "Medicine C", "quantity": 8, "total_price": 4000},
+    ]
+}
