+++ CHEMIST/forms.py
@@ -1,0 +1,40 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, PasswordField, SubmitField, IntegerField, DecimalField
+from wtforms.validators import DataRequired, Email, EqualTo, ValidationError
+from app import mysql
+
+
+class RegisterForm(FlaskForm):
+    name = StringField("Name", validators=[DataRequired()])
+    email = StringField("Email", validators=[DataRequired(), Email()])
+    password = PasswordField("Password", validators=[DataRequired()])
+    confirm_password = PasswordField("Confirm Password", validators=[DataRequired(), EqualTo('password', message='Passwords must match')])
+    submit = SubmitField("Register")
+
+    def validate_email(self, field):
+        cursor = mysql.connection.cursor()
+        cursor.execute("SELECT * FROM admin WHERE email=%s", (field.data,))
+        user = cursor.fetchone()
+        cursor.close()
+        if user:
+            raise ValidationError('Email Already Taken')
+
+class LoginForm(FlaskForm):
+    email = StringField("Email", validators=[DataRequired(), Email()])
+    password = PasswordField("Password", validators=[DataRequired()])
+    submit = SubmitField("Login")
+
+class EditForm(FlaskForm):
+    name = StringField("Name", validators=[DataRequired()])
+    email = StringField("Email", validators=[DataRequired(), Email()])
+    submit = SubmitField("Save Changes")
+
+class AddToCartForm(FlaskForm):
+    quantity = IntegerField("Quantity", validators=[DataRequired()])
+    submit = SubmitField("Add to Cart")
+
+class AddProductForm(FlaskForm):
+    name = StringField('Product Name', validators=[DataRequired()])
+    price = DecimalField('Price', validators=[DataRequired()])
+    description = StringField('Description', validators=[DataRequired()])
+    submit = SubmitField('Add Product')
