+++ requirements.txt
@@ -1,0 +1,5 @@
+Flask==3.0.3
+Flask-WTF==1.2.1
+WTForms==3.1.2
+Flask-MySQLdb==2.0.0
+python-dotenv==1.0.1
